% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare.affection.R
\name{declare.affection}
\alias{declare.affection}
\title{declare.affection}
\usage{
declare.affection(name, oodle = FALSE, ...)
}
\arguments{
\item{name}{A single character string.}

\item{oodle}{A single logical value; should \code{name} be oodle-ized?}

\item{...}{Other arguments to be passed to other methods}
}
\value{
Invisibly returns \code{name}.
}
\description{
Declare affection for \code{name}.
}
\examples{
declare.affection("Alese")
declare.affection("Jeremy", oodle = TRUE, replace.y = TRUE)
}
\seealso{
\code{\link{oodle}}
}
