% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.nontrivial.vector.R
\name{is.nontrivial.vector}
\alias{is.nontrivial.vector}
\title{is.nontrivial.vector}
\usage{
is.nontrivial.vector(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
TRUE or FALSE
}
\description{
Determine if an object is a vector of length > 1.
}
\examples{
## TRUE
is.vector("Hello")
## FALSE
is.nontrivial.vector("Hello")
}
