% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinda.type.R
\name{kinda}
\alias{kinda}
\alias{kinda.type}
\alias{kinda.logical}
\alias{kinda.integer}
\alias{kinda.numeric}
\alias{kinda.character}
\alias{kinda.vector}
\alias{kinda.factor}
\alias{kinda.matrix}
\alias{kinda.array}
\alias{kinda.data.frame}
\alias{kinda.formula}
\alias{kinda.list}
\title{Kinda functions}
\usage{
kinda.type(x, FUN)

kinda.logical(x)

kinda.integer(x)

kinda.numeric(x)

kinda.character(x)

kinda.vector(x)

kinda.factor(x)

kinda.matrix(x)

kinda.array(x)

kinda.data.frame(x)

kinda.formula(x)

kinda.list(x)
}
\arguments{
\item{x}{An R object.}

\item{FUN}{A function of the form as.<type> (e.g. as.logical, as.matrix) that tries to coerce x to <type>.}
}
\description{
kinda.<type>(x) returns whether 'x' can be coerced without error or warning
  to <type> using as.<type>(x).
}
\details{
Be careful using these! Make sure they're doing exactly what you expect, as some functions are pretty lenient in their coercion.
}
