% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{A single logical value; should NAs be removed before proceeding?}
}
\value{
The mode of x. If there are several modes, it returns the smallest one.
}
\description{
Calculate the mode of a vector.
}
\details{
This was adapted from the web.
}
\examples{
Mode(c(1,1,2,2)) # will find the smallest
Mode(c("1", "1", "2", "3"))
Mode(c(TRUE, TRUE, FALSE, FALSE, FALSE, NA))
Mode(factor(c("M", "F", "M")))
}
