% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oodle.R
\name{oodle}
\alias{oodle}
\title{oodle}
\usage{
oodle(name, replace.y = FALSE, ...)
}
\arguments{
\item{name}{A single character string.}

\item{replace.y}{A single logical value; should y's be replaced?}

\item{...}{Other arguments}
}
\value{
A copy of \code{name} with all vowels replaced with 'oodle'.
}
\description{
Replace all vowels (aeiou) in a name with 'oodle'.
}
\examples{
oodle("Brendan Broderick")
oodle("Peter Martin")
oodle("Jeremy Syrjanen", TRUE)
}
