% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowSDs.R
\name{rowSDs}
\alias{rowSDs}
\title{rowSDs}
\usage{
rowSDs(x, na.rm = FALSE)
}
\arguments{
\item{x}{a matrix or data frame, containing numeric, integer, or logical values.}

\item{na.rm}{logical. Should missing values be omitted from calculations?}
}
\value{
A vector with length \code{nrow(x)} containing the standard deviation of each row of x.
}
\description{
Find the standard deviation of rows; analagous to rowMeans.
}
\examples{
myMatrix <- rbind(c(1,2,3,4,5), c(TRUE, FALSE, TRUE, TRUE, FALSE), c(0.3, 0.2, NA, 0.1, 0.1))
rownames(myMatrix) <- c("a", "b", "c")
rowSDs(myMatrix, na.rm = TRUE)
}
